:recycle: refactor: modal, toast 위치 이동 (#257)
