+++ src/components/management/Management.tsx
@@ -1,5 +1,5 @@
 import useFetchManagement from '@/hooks/management/useFetchManagement';
-import useToast from '@/hooks/toast/useToast';
+import useToast from '@/hooks/service/ui/useToast';
 import { submitDetectedOrder } from '@/server/api/supabase/management';
 import useAuthState from '@/shared/store/session';
 import { useEffect } from 'react';
