+++ src/components/menu-item/options/MenuItemFormOption.tsx
@@ -1,4 +1,4 @@
-import { useModal } from '@/hooks/modal/useModal';
+import { useModal } from '@/hooks/service/ui/useModal';
 import useMenuItemStore from '@/shared/store/menu-item';
 import { MenuOptionWithDetail } from '@/types/supabase';
 import { FiAlertCircle } from 'react-icons/fi';
