+++ src/components/menu-item/options/MenuOptionModalInput.tsx
@@ -1,4 +1,4 @@
-import { useModal } from '@/hooks/modal/useModal';
+import { useModal } from '@/hooks/service/ui/useModal';
 import useMenuItemStore from '@/shared/store/menu-item';
 import clsx from 'clsx';
 import React, { useState } from 'react';
